separate error process function for email
